change configuration

+++ tests/Unit/Drivers/LaravelLogDriverTest.php
@@ -16,8 +16,10 @@ class LaravelLogDriverTest extends \Tests\BaseTestCase
      */
     public function it_stores_generic_data()
     {
-        unlink($this->logFileAddress);
-        $this->app['config']->set('laravel-scenario-logger.default', 'laravel-log');
+        if (file_exists($this->logFileAddress)) {
+            unlink($this->logFileAddress);
+        }
+        $this->app['config']->set('laravel-scenario-logger.default_storage_driver', 'laravel_log');
 
         $data = ['a' => 1, 'b' => 2, 'c' => 3];
         app()->make(StorageService::class)->store($data);
@@ -32,7 +34,7 @@ class LaravelLogDriverTest extends \Tests\BaseTestCase
     public function it_stores_modules_data()
     {
         unlink($this->logFileAddress);
-        $this->app['config']->set('laravel-scenario-logger.default', 'laravel-log');
+        $this->app['config']->set('laravel-scenario-logger.default_storage_driver', 'laravel_log');
 
         $data = [
             'services' => [
